+++ FEATURES.md
@@ -89,3 +89,10 @@ for a detailed explanation.
   as well as not restricting their location in the string.
 
   Added in [#4617](https://github.com/emberjs/ember.js/pull/4617)
+
+* `ember-runtime-proxy-mixin`
+
+  Makes the logic behind `Ember.ObjectProxy` into a `Mixin` that can be used when you cannot
+  inherit from `Ember.ObjectProxy` directly.
+
+  Added in [#5156](https://github.com/emberjs/ember.js/pull/5156)
