+++ packages/ember-runtime/lib/main.js
@@ -44,6 +44,7 @@ import {
   Freezable,
   FROZEN_ERROR
 } from "ember-runtime/mixins/freezable";
+import ProxyMixin from "ember-runtime/mixins/proxy";
 
 import Observable from "ember-runtime/mixins/observable";
 import ActionHandler from "ember-runtime/mixins/action_handler";
@@ -167,6 +168,10 @@ Ember.ArrayController = ArrayController;
 Ember.ObjectController = ObjectController;
 Ember.Controller = Controller;
 Ember.ControllerMixin = ControllerMixin;
+
+if (Ember.FEATURES.isEnabled('ember-runtime-proxy-mixin')) {
+  Ember.ProxyMixin = ProxyMixin;
+}
 
 Ember.RSVP = RSVP;
 // END EXPORTS
